# template: speedup lexer