--- a/lib/template/parse/parse.go
+++ b/lib/template/parse/parse.go
@@ -195,7 +195,6 @@ func (t *Tree) recover(errp *error) {
 			panic(e)
 		}
 		if t != nil {
-			t.lex.drain()
 			t.stopParse()
 		}
 		*errp = e.(error)