Speed up the template lexer by ~4x.
```
name                 old time/op    new time/op    delta
Parse/lorem_ipsum-4    2.01µs ± 7%    0.73µs ± 2%  -63.81%  (p=0.000 n=7+8)
Parse/short-4          70.2µs ± 1%    15.2µs ± 1%  -78.39%  (p=0.000 n=7+8)
Parse/medium-4          165µs ± 1%      38µs ± 1%  -77.05%  (p=0.000 n=8+8)
Parse/long-4            517µs ± 1%     117µs ± 5%  -77.31%  (p=0.000 n=8+8)
Parse/very-long-4      1.79ms ± 1%    0.40ms ± 1%  -77.67%  (p=0.000 n=8+8)

name                 old alloc/op   new alloc/op   delta
Parse/lorem_ipsum-4    1.23kB ± 0%    1.16kB ± 0%   -5.56%  (p=0.000 n=8+8)
Parse/short-4          5.48kB ± 0%    5.08kB ± 0%   -7.30%  (p=0.000 n=8+8)
Parse/medium-4         12.7kB ± 0%    12.7kB ± 0%     ~     (all equal)
Parse/long-4           35.3kB ± 0%    34.4kB ± 0%   -2.75%  (p=0.000 n=8+8)
Parse/very-long-4       113kB ± 0%     110kB ± 0%   -2.42%  (p=0.000 n=8+8)

name                 old allocs/op  new allocs/op  delta
Parse/lorem_ipsum-4      12.0 ± 0%      10.0 ± 0%  -16.67%  (p=0.000 n=8+8)
Parse/short-4             109 ± 0%        93 ± 0%  -14.68%  (p=0.000 n=8+8)
Parse/long-4              779 ± 0%       705 ± 0%   -9.50%  (p=0.000 n=8+8)
Parse/very-long-4       2.61k ± 0%     2.35k ± 0%   -9.84%  (p=0.000 n=8+8)
```

Best reviewed commit-by-commit; see associated messages for justification and intermediate benchmark results.

Note that although this is a large diff, a large portion of it is due to benchmark fixtures which may be ignored.